Change interpolation to take max freq instead of average.
Ok so, as far as I can tell the fourier transform we are doing is
called a DFT (Discrete Fourier Transform). The algorithm itself is
called FFT (Fast Fourier Transform) and is a way to compute the DFT.
The DFT is imprecise on two axis:
- time axis (you have to take a rectangular window of data from the
  signal, which has implications because the fourier transform is
  only correct on an infinite signal)
- frequency axis (you're limited by the sampling rate and the size of
  the window of data given)

On top of that, both of these precision problems are inversely related.
This is the eisenberg uncertainty principle: you cannot measure both the
position of a peak precisely (in time) and the frequencies composing
the signal. To increase the precision of the fourier transform, you
need to get an infinite signal. To increase the precision of finding
where a peak is, you need to narrow down the window to a very finite
chunk.

So that means there's theory that proves that we can't do very well.
ON TOP OF THAT, we are in a finite and digital world, which means we
have to deal with a sampling rate which will limit the precision of our
fourier analysis to anything above samplingRate / 2 (but it's not a
step function, the error rate increase as we approach samplingRate / 2).
ON TOP OF _THAT_ the frequencies buckets themselves depend on the
amount of data processed. The more data (compared to the sampling rate)
the more "precision" in the frequency domain you have, BUT humans mostly
care about whole-step Hz and not 1.33333333333...332Hz. So if you give
the algorithm a sample that is 3s (which means the number of points it
contains is 3x the sampling rate), the FFT algorithm will put the
energy into 3 buckets for each whole Hz (0.33333Hz, 0.66666Hz,
0.99999Hz). This is annoying because I don't think you can simply get
the amount of energy at 1Hz by looking at these other buckets. 1Hz is
not the same as 1.3Hz, nor is it the average over 0.9Hz and 1.3Hz as
far as I can tell.

So I simply went with the approach of taking all of these imperfect
buckets (which depend on the size of the signal given) and found the
max amplitude within each whole-Hz bucket. This allows me to say that
at least if there's a peak somewhere, the amount of energy that you
see in the time domain will be "conserved" in the frequency domain
(I think).

There is something called a DTFT (Discrete Time Fourier Transform)
which allows representing the frequency graph in a continuous way that
can be sampled (at exactly the frequencies you care about). This would
be the better way to do this analysis when you actually care about
sub-Hz amplitudes for equipment of sorts. Otherwise this should be fine.